pyproject.toml
[project]
name = "mini-redis-grpc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.7",
]
